Turn received PulseAudio stereo data into mono
Prior to this commit there was a bug in the way the PulseAudio data
source parses the data received from the server, which caused it to
interlace samples of the two received channels and treat them as one
channel. With this commit the data source will average out the two
channels when writing the final buffer.